+++ packages/core-components/src/components/FavoriteToggle/FavoriteToggle.tsx
@@ -63,6 +63,18 @@ export function FavoriteToggleIcon(props: { isFavorite: boolean }) {
 }
 
 /**
+ * Props for the {@link FavoriteToggle} component.
+ *
+ * @public
+ */
+export type FavoriteToggleProps = ComponentProps<typeof IconButton> & {
+  id: string;
+  title: string;
+  isFavorite: boolean;
+  onToggle: (value: boolean) => void;
+};
+
+/**
  * Toggle encapsulating logic for marking something as favorite,
  * primarily used in various instances of entity lists and cards but can be used elsewhere.
  *
@@ -70,14 +82,7 @@ export function FavoriteToggleIcon(props: { isFavorite: boolean }) {
  *
  * @public
  */
-export function FavoriteToggle(
-  props: ComponentProps<typeof IconButton> & {
-    id: string;
-    title: string;
-    isFavorite: boolean;
-    onToggle: (value: boolean) => void;
-  },
-) {
+export function FavoriteToggle(props: FavoriteToggleProps) {
   const {
     id,
     title,