+++ packages/frontend-app-api/src/tree/instantiateAppNodeTree.test.ts
@@ -15,9 +15,9 @@
  */
 
 import {
-  AnyExtensionDataRef,
   AppNode,
   Extension,
+  ExtensionDataRef,
   ExtensionInput,
   PortableSchema,
   ResolvedExtensionInput,
@@ -56,7 +56,6 @@ function makeSpec<TConfig, TConfigInput>(
     attachTo: extension.attachTo,
     disabled: extension.disabled,
     extension: extension as Extension<unknown, unknown>,
-    source: undefined,
     plugin: undefined,
     ...spec,
   };
@@ -90,7 +89,7 @@ function makeInstanceWithId<TConfig, TConfigInput>(
 }
 
 function createV1ExtensionInput(
-  extensionData: Record<string, AnyExtensionDataRef>,
+  extensionData: Record<string, ExtensionDataRef>,
   options: { singleton?: boolean; optional?: boolean } = {},
 ) {
   return {
@@ -108,7 +107,7 @@ function createV1Extension(opts: {
   name?: string;
   attachTo?: { id: string; input: string };
   inputs?: Record<string, ReturnType<typeof createV1ExtensionInput>>;
-  output: Record<string, AnyExtensionDataRef>;
+  output: Record<string, ExtensionDataRef>;
   configSchema?: PortableSchema<any, any>;
   factory: (ctx: { inputs: any; config: any }) => any;
 }): Extension<any, any> {