+++ packages/frontend-app-api/src/wiring/createSpecializedApp.tsx
@@ -25,7 +25,6 @@ import {
   SubRouteRef,
   AnyRouteRefParams,
   RouteFunc,
-  RouteResolutionApiResolveOptions,
   RouteResolutionApi,
   createApiFactory,
   routeResolutionApiRef,
@@ -81,6 +80,7 @@ import {
   createPluginInfoAttacher,
   FrontendPluginInfoResolver,
 } from './createPluginInfoAttacher';
+import { createRouteAliasResolver } from '../routing/RouteAliasResolver';
 
 function deduplicateFeatures(
   allFeatures: FrontendFeature[],
@@ -169,7 +169,7 @@ class RouteResolutionApiProxy implements RouteResolutionApi {
       | RouteRef<TParams>
       | SubRouteRef<TParams>
       | ExternalRouteRef<TParams>,
-    options?: RouteResolutionApiResolveOptions,
+    options?: { sourcePath?: string },
   ): RouteFunc<TParams> | undefined {
     if (!this.#delegate) {
       throw new Error(
@@ -187,6 +187,7 @@ class RouteResolutionApiProxy implements RouteResolutionApi {
       routeInfo.routeObjects,
       this.routeBindings,
       this.appBasePath,
+      routeInfo.routeAliasResolver,
     );
     this.#routeObjects = routeInfo.routeObjects;
 
@@ -199,13 +200,11 @@ class RouteResolutionApiProxy implements RouteResolutionApi {
 }
 
 /**
- * Creates an empty app without any default features. This is a low-level API is
- * intended for use in tests or specialized setups. Typically you want to use
- * `createApp` from `@backstage/frontend-defaults` instead.
+ * Options for `createSpecializedApp`.
  *
  * @public
  */
-export function createSpecializedApp(options?: {
+export type CreateSpecializedAppOptions = {
   features?: FrontendFeature[];
   config?: ConfigApi;
   bindRoutes?(context: { bind: CreateAppRouteBinder }): void;
@@ -215,7 +214,19 @@ export function createSpecializedApp(options?: {
     | ExtensionFactoryMiddleware[];
   flags?: { allowUnknownExtensionConfig?: boolean };
   pluginInfoResolver?: FrontendPluginInfoResolver;
-}): { apis: ApiHolder; tree: AppTree } {
+};
+
+/**
+ * Creates an empty app without any default features. This is a low-level API is
+ * intended for use in tests or specialized setups. Typically you want to use
+ * `createApp` from `@backstage/frontend-defaults` instead.
+ *
+ * @public
+ */
+export function createSpecializedApp(options?: CreateSpecializedAppOptions): {
+  apis: ApiHolder;
+  tree: AppTree;
+} {
   const config = options?.config ?? new ConfigReader({}, 'empty-config');
   const features = deduplicateFeatures(options?.features ?? []).map(
     createPluginInfoAttacher(config, options?.pluginInfoResolver),
@@ -237,12 +248,10 @@ export function createSpecializedApp(options?: {
   const factories = createApiFactories({ tree });
   const appBasePath = getBasePath(config);
   const appTreeApi = new AppTreeApiProxy(tree, appBasePath);
+
+  const routeRefsById = collectRouteIds(features);
   const routeResolutionApi = new RouteResolutionApiProxy(
-    resolveRouteBindings(
-      options?.bindRoutes,
-      config,
-      collectRouteIds(features),
-    ),
+    resolveRouteBindings(options?.bindRoutes, config, routeRefsById),
     appBasePath,
   );
 
@@ -288,7 +297,10 @@ export function createSpecializedApp(options?: {
     mergeExtensionFactoryMiddleware(options?.extensionFactoryMiddleware),
   );
 
-  const routeInfo = extractRouteInfoFromAppNode(tree.root);
+  const routeInfo = extractRouteInfoFromAppNode(
+    tree.root,
+    createRouteAliasResolver(routeRefsById),
+  );
 
   routeResolutionApi.initialize(routeInfo);
   appTreeApi.initialize(routeInfo);