+++ packages/frontend-app-api/src/wiring/types.ts
@@ -14,17 +14,13 @@
  * limitations under the License.
  */
 import { RouteRef } from '@backstage/frontend-plugin-api';
-import { FrontendFeature as PluginApiFrontendFeature } from '@backstage/frontend-plugin-api';
 import { BackstageRouteObject } from '../routing/types';
-
-/** @public
- * @deprecated Use {@link @backstage/frontend-plugin-api#FrontendFeature} instead.
- */
-export type FrontendFeature = PluginApiFrontendFeature;
+import { RouteAliasResolver } from '../routing/RouteAliasResolver';
 
 /** @internal */
 export type RouteInfo = {
   routePaths: Map<RouteRef, string>;
   routeParents: Map<RouteRef, RouteRef | undefined>;
   routeObjects: BackstageRouteObject[];
+  routeAliasResolver: RouteAliasResolver;
 };