+++ packages/frontend-plugin-api/src/blueprints/index.ts
@@ -14,6 +14,10 @@
  * limitations under the License.
  */
 
+export {
+  AnalyticsImplementationBlueprint,
+  type AnalyticsImplementationFactory,
+} from './AnalyticsImplementationBlueprint';
 export { ApiBlueprint } from './ApiBlueprint';
 export { AppRootElementBlueprint } from './AppRootElementBlueprint';
 export { AppRootWrapperBlueprint } from './AppRootWrapperBlueprint';