+++ packages/frontend-plugin-api/src/wiring/index.ts
@@ -55,8 +55,6 @@ export {
 } from './createFrontendFeatureLoader';
 export { type Extension } from './resolveExtensionDefinition';
 export {
-  type AnyRoutes,
-  type AnyExternalRoutes,
   type ExtensionDataContainer,
   type FeatureFlagConfig,
   type ExtensionFactoryMiddleware,
@@ -67,7 +65,7 @@ export {
   type ExtensionBlueprint,
   type ExtensionBlueprintParameters,
   type ExtensionBlueprintParams,
-  type ExtensionBlueprintParamsDefiner,
+  type ExtensionBlueprintDefineParams,
   createExtensionBlueprint,
   createExtensionBlueprintParams,
 } from './createExtensionBlueprint';