+++ plugins/catalog-unprocessed-entities/src/alpha/plugin.tsx
@@ -36,8 +36,8 @@ import { rootRouteRef } from '../routes';
 
 /** @alpha */
 export const catalogUnprocessedEntitiesApi = ApiBlueprint.make({
-  params: define =>
-    define({
+  params: defineParams =>
+    defineParams({
       api: catalogUnprocessedEntitiesApiRef,
       deps: {
         discoveryApi: discoveryApiRef,
@@ -51,7 +51,7 @@ export const catalogUnprocessedEntitiesApi = ApiBlueprint.make({
 /** @alpha */
 export const catalogUnprocessedEntitiesPage = PageBlueprint.make({
   params: {
-    defaultPath: '/catalog-unprocessed-entities',
+    path: '/catalog-unprocessed-entities',
     routeRef: convertLegacyRouteRef(rootRouteRef),
     loader: () =>
       import('../components/UnprocessedEntities').then(m =>