+++ plugins/search-react/src/alpha/blueprints/SearchResultListItemBlueprint.test.tsx
@@ -91,7 +91,7 @@ describe('SearchResultListItemBlueprint', () => {
       },
       factory(originalFactory, { inputs }) {
         return originalFactory({
-          defaultPath: '/',
+          path: '/',
           loader: async () => {
             const items = inputs.items.map(i =>
               i.get(searchResultListItemDataRef),